Custom rate limiter for the Broadcast
