+++ config/twilio.js
@@ -5,7 +5,7 @@ const envPrefix = useTestCreds ? 'TWILIO_API_TEST_' : 'TWILIO_API_';
 
 const config = {
   useTestCreds,
-  accountSid: process.env[`${envPrefix}ACCOUNT_SID`] || 'account_sid',
+  accountSid: process.env[`${envPrefix}ACCOUNT_SID`] || 'AC-account_sid',
   authToken: process.env[`${envPrefix}AUTH_TOKEN`] || 'totallysecret',
   from: process.env[`${envPrefix}FROM`] || '+15005550006',
 };
