+++ test/integration/web/controllers/WebHooksWebController.test.js
@@ -128,40 +128,6 @@ test('POST /api/v1/webhooks/twilio-sms-inbound should publish message to twilio-
 /**
  * POST /api/v1/webhooks/customerio-sms-broadcast
  */
-test('POST /api/v1/webhooks/customerio-sms-broadcast should publish message to customerio-sms-broadcast-relay queue', async (t) => {
-  const broadcastId = chance.word();
-  const data = MessageFactoryHelper.getValidCustomerBroadcastData(broadcastId);
-
-  const res = await t.context.supertest.post('/api/v1/webhooks/customerio-sms-broadcast')
-    .set('Content-Type', 'application/json')
-    .auth(t.context.config.app.auth.name, t.context.config.app.auth.password)
-    .send(data);
-
-  // Ensure Customer.io compatible response.
-  res.status.should.be.equal(201);
-
-  // Check response to be json
-  res.header.should.have.property('content-type');
-  res.header['content-type'].should.match(/json/);
-
-  // Check response.
-  res.body.should.have.property('ok', true);
-
-  // Check that the message is queued.
-  const rabbit = new RabbitManagement(t.context.config.amqpManagement);
-  const messages = await rabbit.getMessagesFrom('customerio-sms-broadcast-relay', 1, false);
-  messages.should.be.an('array').and.to.have.lengthOf(1);
-
-  messages[0].should.have.property('payload');
-  const payload = messages[0].payload;
-  const messageData = JSON.parse(payload);
-  messageData.should.have.property('data');
-  messageData.data.should.be.eql(data);
-});
-
-/**
- * POST /api/v1/webhooks/customerio-sms-broadcast
- */
 test('POST /api/v1/webhooks/customerio-sms-broadcast validates incoming payload', async (t) => {
   const broadcastId = chance.word();
   const data = MessageFactoryHelper.getValidCustomerBroadcastData(broadcastId);
